+++ ospac/__init__.py
@@ -14,7 +14,8 @@ try:
 except PackageNotFoundError:  # running from a source tree that was never installed
     __version__ = "0.0.0.dev0"
 
-from ospac.aliases import license_aliases, license_ambiguous, license_never_resolve
+from ospac.aliases import (LicenseResolution, license_aliases, license_ambiguous,
+                           license_never_resolve, resolve_license)
 from ospac.dataset import DATA_SCHEMA_VERSION, DataVersion, data_version
 from ospac.runtime.engine import PolicyRuntime
 from ospac.models.license import License
@@ -23,7 +24,9 @@ from ospac.models.compliance import ComplianceResult
 
 __all__ = [
     "PolicyRuntime",
+    "LicenseResolution",
     "license_aliases",
+    "resolve_license",
     "license_ambiguous",
     "license_never_resolve",
     "data_version",